Allow client to approve candidate for a job
