+++ client/src/pages/JobPosting.js
@@ -1,113 +1,92 @@
 import React, { Component } from 'react';
 
-import skills from '../utils/skills';
-
-class JobPosting extends Component {
+class Posts extends Component {
   constructor(props) {
     super(props);
 
     this.state = {
-      title: '',
-      description: '',
-      skills: []
+      jobs: []
     };
+    console.log('helllo')
 
-    this.handleChange = this.handleChange.bind(this);
-    this.handleSubmit = this.handleSubmit.bind(this);
+    this.handleClick = this.handleClick.bind(this);
   }
 
-  componentDidMount() {
-    if (!localStorage.getItem('token') || localStorage.getItem('token') === 'undefined') {
-      return this.props.history.push('/login');
-    }
-  }
-
-  handleChange(e) {
-    const jobSkills = [];
-    if (e.target.type === 'select-multiple') {
-      const options = e.target.options;
-      for (let i = 0; i < options.length; i++) {
-        if (options[i].selected) {
-          jobSkills.push(options[i].value);
-        }
-      }
-    }
-    
-    if (this.state.error) this.setState({ error: false });
-    const { name, value } = e.target;
-    this.setState(() => ({
-      [name]: jobSkills.length ? jobSkills : value
-    }));
-  }
-
-  async handleSubmit(e) {
-    e.preventDefault();
-    const data = await this.createJob();
-    if (!data.error) {
-      return this.props.history.push('/');
-    }
-    this.setState({ error: true });
-  }
-
-  async createJob() {
-    const { title, description, skills } = this.state;
+  async componentDidMount() {
+    this.ref = window.document.getElementById('alert');
     try {
-      const res = await fetch('/api/v1/jobs', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          'x-access-token': localStorage.getItem('token')
-        },
-        body: JSON.stringify({ title, description, skills })
+      const token = localStorage.getItem('token');
+      if (!token || token === 'undefined') return this.props.history.push('/');
+      console.log('got here?')
+      const res = await fetch('/api/v1/jobs', { headers: { 'x-access-token': token } });
+      const data = await res.json();
+      this.setState({
+        jobs: data
       });
-      return res.json();
     } catch (error) {
       console.log(error);
     }
   }
 
+  async handleClick(job) {
+    try {
+      const res = await fetch('/api/v1/applicants', {
+        method: 'POST',
+          headers: {
+            'Content-Type': 'application/json',
+            'x-access-token': localStorage.getItem('token')
+          },
+          body: JSON.stringify({ job_id: job.job_id })
+      });
+
+      const data = await res.json();
+      if (data.error || data.errors) {
+        return this.handleError();
+      }
+      return this.props.history.push('/');
+    } catch (error) {
+      console.log(error);
+    }
+  }
+
+  handleError() {
+    window.scrollTo(0, document.getElementById('alert').offsetTop - 20);
+    this.setState({ error: true });
+    return;
+  }
+
   render() {
-    const skillOptions = skills.map((item) => (
-      <option key={item.id} value={item.name}>
-        {item.name}
-      </option>
+    const jobs = this.state.jobs.map((job) => (
+      <div key={job.job_id} className='card text-center' style={{ marginTop: '2rem' }}>
+          <div className='card-header'>
+          {job.title}
+          </div>
+          <div className='card-body'>
+            <p className='card-text'>{job.description}</p>
+            <p><b>Required Skills</b>: {job.skills.join(', ')}</p>
+            <button className='btn btn-secondary' type='button' style={{ width: '7rem' }} onClick={() => this.handleClick(job)}>Apply</button>
+
+          </div>
+          <div className='card-footer text-muted'>
+            {new Date(job.updated_at).toLocaleDateString()}
+          </div>
+        </div>
     ));
     return (
-      <div className='container' style={{
-        padding: '1.5rem',
-        margin: '0 auto',
-        borderWidth: '2rem',
-        width: '40rem'
-      }}>
-        <h3>Post Job</h3>
-        {this.state.error && <div className='alert alert-danger' role='alert'>
-          An error occurred with your application :(
-        </div>}
-        <form className='container'>
-          <div className='form-group'>
-            <label htmlFor='title'>Job Title</label>
-            <input type='text' name='title' id='title' className='form-control form-control-sm' value={this.state.title} onChange={this.handleChange} />
+      <div className='container'>
+        <h3>Jobs</h3>
+        {
+          <div id='alert'>
+          { this.state.error &&
+          <div className='alert alert-danger' role='alert'>
+            An error occurred with your application :(
+            </div> }
           </div>
-          <div className='form-group'>
-            <label htmlFor='des'>Skills</label>
-            <input readOnly type='text' name='skills' id='skills' className='form-control form-control-sm' value={this.state.skills} onChange={this.handleChange} />
-          </div>
-          <div className='form-group'>
-            <select multiple className='form-control form-control-sm' onChange={this.handleChange} name='skills'>
-              {skillOptions}
-            </select>
-          </div>
-          <div className='form-group'>
-            <label htmlFor='description'>Description</label>
-            <textarea className='form-control form-control-sm' id='description' name='description' value={this.state.description} onChange={this.handleChange} rows='3' />
-          </div>
-          <div className='form-group'>
-            <button className='btn btn-primary btn-lg' type='submit' onClick={this.handleSubmit}>Post</button>
-          </div>
-        </form>
+        }
+        {jobs}
       </div>
-    )
+    );
   }
 }
 
-export default JobPosting;
+export default Posts;
